requirements.txt
# requirements.txt
requests
yfinance
pandas
python-dotenv
